fixed all routing and backend errors associated with rating averages
