+++ src/services/ratingRoutes.ts
@@ -1,21 +1,29 @@
-import { RatingModel } from './RatingModel.js';
+import { RatingModel } from './ratingModel.js';
 
-export const ratingRoutes = (app: any) => {
-  app.post('/api/ratings', async (req: any, res: any) => {
-    const ratingData = await req.json();
-    const newRating = await RatingModel.create(ratingData);
-    return res.json(newRating);
-  });
+export const ratingRoutes = (req: Request) => {
+  const url = new URL(req.url);
+  const method = req.method;
 
-  app.get('/api/ratings/:restaurantName', async (req: any, res: any) => {
-    const restaurantName = req.params.restaurantName; // Use name instead of ID
-    const ratings = await RatingModel.getRatingsByRestaurant(restaurantName);
-    return res.json(ratings);
-  });
+  if (url.pathname === '/api/ratings' && method === 'POST') {
+    return req.json().then(async (ratingData) => {
+      const newRating = await RatingModel.create(ratingData);
+      return new Response(JSON.stringify(newRating), { status: 200 });
+    });
+  }
 
-  app.get('/api/ratings/average/:restaurantName', async (req: any, res: any) => {
-    const restaurantName = req.params.restaurantName; // Use name instead of ID
-    const averageRating = await RatingModel.getAverageRating(restaurantName);
-    return res.json({ averageRating });
-  });
+  if (url.pathname.startsWith('/api/ratings/average/') && method === 'GET') {
+    const restaurantName = url.pathname.split('/').pop();
+    return RatingModel.getAverageRating(restaurantName || '').then((averageRating) =>
+      new Response(JSON.stringify({ averageRating }), { status: 200 })
+    );
+  }
+
+  if (url.pathname.startsWith('/api/ratings/') && method === 'GET') {
+    const restaurantName = url.pathname.split('/').pop();
+    return RatingModel.getRatingsByRestaurant(restaurantName || '').then((ratings) =>
+      new Response(JSON.stringify(ratings), { status: 200 })
+    );
+  }
+
+  return Promise.resolve(new Response('Not Found', { status: 404 }));
 };
