+++ src/util/server.ts
@@ -1,12 +1,10 @@
-import { RatingModel } from './services/RatingModel.js';
-import { ratingRoutes } from './services/ratingRoutes.ts';
+import { ratingRoutes } from '../services/ratingRoutes';
 
 const server = Bun.serve({
   port: 3000,
-  fetch(req) {
-    // Handle routes using ratingRoutes
-    const url = new URL(req.url);
-    if (url.pathname.startsWith('/api/ratings')) {
+  async fetch(req) {
+    // Delegate to ratingRoutes for handling API requests
+    if (req.url.startsWith('/api/ratings')) {
       return ratingRoutes(req);
     }
     return new Response('Not Found', { status: 404 });
@@ -14,4 +12,3 @@ const server = Bun.serve({
 });
 
 console.log(`Server running on http://localhost:${server.port}`);
-
